MESS MANAGEMENT - Copy/backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
email-validator
asyncpg
sqlalchemy
